rename gan module name to mnist
for some reason isort detects it correctly when its named
gan incorrectly when its named anything else. but it'd be
consistent with other modules
