+++ examples/gan/gan_mnist_pl/mnist.py
